+++ src/Raven.CodeAnalysis/BoundTree/BoundBaseExpression.cs
@@ -1,0 +1,9 @@
+namespace Raven.CodeAnalysis;
+
+internal partial class BoundBaseExpression : BoundExpression
+{
+    public BoundBaseExpression(ITypeSymbol type, BoundExpressionReason reason = BoundExpressionReason.None)
+        : base(type, type, reason)
+    {
+    }
+}